Added regularity of backups.
Not every project needs to be exported / backed up at the same 
regularity, as some projects rarely see progress.
But it would be good to have just one config file for various different 
regularities of backup.

Now only the projects with the selected regularity will be backed up.
Had to change the format of the config.yaml file to achieve this.
Now each project has a name and a regularity.
The name is as before, the regularity is a list of strings.
If any of those strings matches the new command line argument passed 
with -r, then that project will be back up.
Otherwise, that project will not be backed up.